move all errors into dedicated files
