+++ app/controllers/concerns/shopify_app/ensure_authenticated_links.rb
@@ -26,7 +26,7 @@ module ShopifyApp
 
     def redirect_to_splash_page
       redirect_to(splash_page)
-    rescue ShopifyApp::LoginProtection::ShopifyDomainNotFound => error
+    rescue ::ShopifyApp::LoginProtection::ShopifyDomainNotFound => error
       Rails.logger.warn("[ShopifyApp::EnsureAuthenticatedLinks] Redirecting to login: [#{error.class}] "\
         "Could not determine current shop domain")
       redirect_to(ShopifyApp.configuration.login_url)
