+++ lib/shopify_app/controller_concerns/ensure_billing.rb
@@ -2,17 +2,6 @@
 
 module ShopifyApp
   module EnsureBilling
-    class BillingError < StandardError
-      attr_accessor :message
-      attr_accessor :errors
-
-      def initialize(message, errors)
-        super
-        @message = message
-        @errors = errors
-      end
-    end
-
     extend ActiveSupport::Concern
 
     RECURRING_INTERVALS = [BillingConfiguration::INTERVAL_EVERY_30_DAYS, BillingConfiguration::INTERVAL_ANNUAL]
