+++ lib/shopify_app/errors/access_scopes/user_strategy/invalid_input.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  module AccessScopes
+    class UserStrategy
+      class InvalidInput < StandardError; end
+    end
+  end
+end
