+++ lib/shopify_app/errors/in_memory_session_store/environment_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class InMemorySessionStore
+    class EnvironmentError < StandardError; end
+  end
+end
