+++ lib/shopify_app/errors/jwt/invalid_audience_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class JWT
+    class InvalidAudienceError < StandardError; end
+  end
+end
