+++ lib/shopify_app/errors/jwt/invalid_destination_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class JWT
+    class InvalidDestinationError < StandardError; end
+  end
+end
