+++ lib/shopify_app/errors/jwt/mismatched_hosts_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class JWT
+    class MismatchedHostsError < StandardError; end
+  end
+end
