+++ lib/shopify_app/errors/login_protection/shopify_domain_not_found.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  module LoginProtection
+    class ShopifyDomainNotFound < StandardError; end
+  end
+end
