+++ lib/shopify_app/errors/login_protection/shopify_host_not_found.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  module LoginProtection
+    class ShopifyHostNotFound < StandardError; end
+  end
+end
