+++ lib/shopify_app/errors/missing_webhook_job_error.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class MissingWebhookJobError < StandardError; end
+end
