+++ lib/shopify_app/errors/scripttags_manager/creation_failed.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class ScripttagsManager
+    class CreationFailed < StandardError; end
+  end
+end
