+++ lib/shopify_app/errors/session_repository/configuration_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ShopifyApp
+  class SessionRepository
+    class ConfigurationError < StandardError; end
+  end
+end
