+++ lib/shopify_app/session/jwt.rb
@@ -2,12 +2,6 @@
 
 module ShopifyApp
   class JWT
-    class InvalidDestinationError < StandardError; end
-
-    class MismatchedHostsError < StandardError; end
-
-    class InvalidAudienceError < StandardError; end
-
     WARN_EXCEPTIONS = [
       ::JWT::DecodeError,
       ::JWT::ExpiredSignature,
